[backport] Remove nodegit in favor of shell
We were already shelling to the git cli to do the `git apply` command
since libgit2 does not have support for it, so we're already saddled
with the baggage of shelling out to a child process. The git cli is
considerably higher level than nodegit as well, which makes it ideal for
the types of automation we're doing with backporting.