[backport] Allow chaining backports across PRs
You can now call the backport task on a backport PR, and it'll use the
diff from that PR rather than the original to create new backports for
the given targets. There's also an indication in the pull request
message that it was generated through a different backport's diff.

Other than that, the new backport pull request is essentially identical
to the original, so when auditing the backports on any given PR, all
backports will appear identically.