[backport] Queue backports on the same repo
The backport command uses the GitQueue utility to register the backport
command for the current repo. This ensures that backporting will only
run when there are no other commands running on that repo, which means
we should no longer need to worry about race conditions when running
multiple backports at the same time.